move help cmd tests to own module
